+++ Source/nt.Client/Features/WebThree/Components/NftTemplates/BillOfLading/BillOfLadingTemplate.razor
@@ -1,6 +1,6 @@
 ﻿@inherits BillOfLadingTemplateModel
 
-<div id="BillOfLadingTemplateComponent" class="container-fluid justify-content-around row align-items-center pt-3">
+<div id="BillOfLadingTemplateComBolnent" class="container-fluid justify-content-around row align-items-center pt-3">
 
   <div class="bg-info">
     <h3>@FormData.Title</h3>
@@ -9,53 +9,53 @@
         <ValidationSummary />*@
       <div class="form-row mb-3 p-2">
         <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-          <label for="PoApprover" class="font-weight-bold">Approver</label>
-          <input id="Approver" class="form-control" placeholder="Approver" type="text" @bind="@FormData.Carrier" />
+          <label for="BolCarrier" class="font-weight-bold">Carrier</label>
+          <input id="Carrier" class="form-control" placeholder="Carrier" type="text" @bind="@FormData.Carrier" />
         </div>
         <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-          <label for="PoDepartment" class="font-weight-bold">Department:</label>
-          <input id="Department" type="text" class="form-control" @bind="@FormData.Consignee" />
+          <label for="BolConsignee" class="font-weight-bold">Consignee:</label>
+          <input id="Consignee" type="text" class="form-control" @bind="@FormData.Consignee" />
         </div>
         <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-          <label for="PoDeliverDate" class="font-weight-bold">Deliver Date</label>
-          <input id="DeliveryDate" type="text" class="form-control" @bind="@FormData.Date" />
+          <label for="BolDate" class="font-weight-bold">Date</label>
+          <input id="Date" type="date" class="form-control" @bind="@FormData.Date" />
         </div>
         <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-          <label for="PoRequester" class="font-weight-bold">Requester:</label>
-          <input id="Requester" type="text" class="form-control" @bind="@FormData.Details" />
+          <label for="BolDetails" class="font-weight-bold">Details:</label>
+          <input id="Details" type="text" class="form-control" @bind="@FormData.Details" />
         </div>
         <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-          <label for="PoNotes" class="font-weight-bold">Requester:Notes:</label>
-          <textarea id="Notes" type="text" class="form-control" @bind="@FormData.Instructions" />
+          <label for="BolInstructions" class="font-weight-bold">Instructions:</label>
+          <textarea id="Instructions" type="text" class="form-control" @bind="@FormData.Instructions" />
         </div>
       </div>
       <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-        <label for="PoItem_Name" class="font-weight-bold">Item_Name</label>
-        <input id="Item_Name" type="text" class="form-control" @bind="@FormData.ItemCode" />
+        <label for="BolItemCode" class="font-weight-bold">ItemCode</label>
+        <input id="ItemCode" type="text" class="form-control" @bind="@FormData.ItemCode" />
       </div>
       <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-        <label for="PoItem_Price" class="font-weight-bold">Item Price:</label>
-        <input id="Item_Price" type="text" class="form-control" @bind="@FormData.Lading" />
+        <label for="BolLading" class="font-weight-bold">Lading:</label>
+        <input id="Lading" type="text" class="form-control" @bind="@FormData.Lading" />
       </div>
       <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-        <label for="PoItem_Qty" class="font-weight-bold">Item Qty:</label>
-        <input id="Item_Qty" type="text" class="form-control" @bind="@FormData.Order" />
+        <label for="BolOrder" class="font-weight-bold">Order:</label>
+        <input id="Order" type="text" class="form-control" @bind="@FormData.Order" />
       </div>
       <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-        <label for="PoItem_Code" class="font-weight-bold">Item Code:</label>
-        <input id="Item_Code" type="text" class="form-control" @bind="@FormData.PO" />
+        <label for="BolPO" class="font-weight-bold">Purchase Order:</label>
+        <input id="BolPO" type="text" class="form-control" @bind="@FormData.PO" />
       </div>
       <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-        <label for="PoItem_Discount" class="font-weight-bold">Item_Discount:</label>
-        <input id="Item_Discount" type="text" class="form-control" @bind="@FormData.Quantity" />
+        <label for="Quantity" class="font-weight-bold">Quantity:</label>
+        <input id="Quantity" type="text" class="form-control" @bind="@FormData.Quantity" />
       </div>
       <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-        <label for="PoItem_Label" class="font-weight-bold">Item Total:</label>
-        <input id="Item_Total" type="text" class="form-control" @bind=@FormData.Shipper />
+        <label for="Shipper" class="font-weight-bold">Shipper:</label>
+        <input id="Shipper" type="text" class="form-control" @bind=@FormData.Shipper />
       </div>
       <div class="col-6 col-sm-5 col-lg-3 col-xl-2 text-center m-2">
-        <label for="PoItem_Label" class="font-weight-bold">Item Total:</label>
-        <input id="Item_Total" type="text" class="form-control" @bind=@FormData.Weight />
+        <label for="BolWeight" class="font-weight-bold">Weight:</label>
+        <input id="Weight" type="text" class="form-control" @bind=@FormData.Weight />
       </div>
       <div class="col-6 col-sm-5 col-lg-4 col-xl-3 text-center m-2">
         <label for="AdditionalInformation" class="font-weight-bold">Additional Information:</label>
@@ -64,6 +64,6 @@
     </div>
   </div>
   <button type="button" class="btn btn-primary" @onclick=@SendDataToState>
-    Send 
+   Confirm Values
   </button>
 </div>
